+++ SadConsole/GameHost.cs
@@ -232,7 +232,8 @@ namespace SadConsole
 
             if (Screen == StartingConsole)
                 throw new Exception($"{nameof(Screen)} must be reassigned to a new instance before this method can be called.");
-
+            
+            StartingConsole.IsFocused = false;
             StartingConsole.Dispose();
             StartingConsole = null;
         }