+++ Samples/MainSample/CustomConsoles/MultipleCursors.cs
@@ -15,14 +15,14 @@ namespace FeatureDemo.CustomConsoles
         {
             CursorGreen = new SadConsole.Components.Cursor()
             {
-                CursorRenderCell = new ColoredGlyph(Color.Green, Color.Transparent, SadConsole.Components.Cursor.DefaultCursorGlyph),
+                CursorRenderCell = new ColoredGlyph(Color.Green, Color.Transparent, SadConsole.Components.Cursor.DefaultCursorGlyph).ToState(),
                 PrintAppearance = new ColoredGlyph(Color.Green, Color.Transparent, SadConsole.Components.Cursor.DefaultCursorGlyph),
                 IsEnabled = true,
                 ApplyCursorEffect = true
             };
             CursorYellow = new SadConsole.Components.Cursor()
             {
-                CursorRenderCell = new ColoredGlyph(Color.Yellow, Color.Transparent, SadConsole.Components.Cursor.DefaultCursorGlyph),
+                CursorRenderCell = new ColoredGlyph(Color.Yellow, Color.Transparent, SadConsole.Components.Cursor.DefaultCursorGlyph).ToState(),
                 PrintAppearance = new ColoredGlyph(Color.Yellow, Color.Transparent, SadConsole.Components.Cursor.DefaultCursorGlyph),
                 PrintOnlyCharacterData = false,
                 IsEnabled = false,
@@ -30,7 +30,7 @@ namespace FeatureDemo.CustomConsoles
             };
             CursorPurple = new SadConsole.Components.Cursor()
             {
-                CursorRenderCell = new ColoredGlyph(Color.Purple, Color.Transparent, SadConsole.Components.Cursor.DefaultCursorGlyph),
+                CursorRenderCell = new ColoredGlyph(Color.Purple, Color.Transparent, SadConsole.Components.Cursor.DefaultCursorGlyph).ToState(),
                 PrintAppearance = new ColoredGlyph(Color.Purple, Color.Transparent, SadConsole.Components.Cursor.DefaultCursorGlyph),
                 PrintOnlyCharacterData = false,
                 IsEnabled = false,